backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
psycopg[binary]>=3.2.0
celery==5.3.6
redis==5.0.3
django-cors-headers==4.3.1
python-dotenv==1.0.1
pytest==8.1.1
pytest-django==4.8.0
factory-boy==3.3.0
django-fsm==2.8.1
uritemplate==4.1.1
